requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
httpx>=0.27.0
bolt11>=2.1.0
